+++ src/index.tsx
@@ -1,6 +1,7 @@
+import CheckboxMini from "./CheckboxMini";
 import Checkbox from "./Checkbox";
 import Radio from "./Radio";
 import CheckboxGroup from "./CheckboxGroup";
 import RadioGroup from "./RadioGroup";
 
-export { Checkbox, Radio, CheckboxGroup, RadioGroup };
+export { CheckboxMini, Checkbox, Radio, CheckboxGroup, RadioGroup };